+++ src/snowflake/snowpark/__init__.py
@@ -53,7 +53,7 @@ from snowflake.snowpark.dataframe_reader import DataFrameReader
 from snowflake.snowpark.dataframe_stat_functions import DataFrameStatFunctions
 from snowflake.snowpark.dataframe_writer import DataFrameWriter
 from snowflake.snowpark.file_operation import FileOperation, GetResult, PutResult
-from snowflake.snowpark.open_telemetry_prototype import open_telemetry
+from snowflake.snowpark._internal.open_telemetry_prototype import open_telemetry
 from snowflake.snowpark.query_history import QueryHistory, QueryRecord
 from snowflake.snowpark.relational_grouped_dataframe import (
     GroupingSets,
